requirements.txt
torch
transformers
Pillow
tqdm
matplotlib
numpy
